change classes ans structure of project
