--- registration-app/src/main/java/com/registrationapp/RegistrationAppApplication.java
@@ -1,15 +1,0 @@
-package com.registrationapp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RegistrationAppApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(RegistrationAppApplication.class, args);
-    }
-
-
-
-}
