--- registration-app/src/main/java/com/registrationapp/repository/AttendeeRepository.java
@@ -1,13 +1,0 @@
-package com.registrationapp.repository;
-
-
-import com.registrationapp.model.Attendee;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-
-@Repository
-@Transactional
-public interface AttendeeRepository extends JpaRepository<Attendee, Long> {
-}
