--- registration-app/src/main/java/com/registrationapp/repository/DiscountCodeRepository.java
@@ -1,16 +1,0 @@
-package com.registrationapp.repository;
-
-
-import com.registrationapp.model.DiscountCode;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-import java.util.Optional;
-
-@Repository
-@Transactional
-public interface DiscountCodeRepository extends JpaRepository<DiscountCode, Long> {
-
-    Optional<DiscountCode> findByCode(String code);
-}
