--- registration-app/src/main/java/com/registrationapp/repository/PricingCategoryRepository.java
@@ -1,23 +1,0 @@
-package com.registrationapp.repository;
-
-
-import com.registrationapp.model.PricingCategory;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-import java.time.LocalDate;
-import java.util.Optional;
-
-@Repository
-@Transactional
-public interface PricingCategoryRepository extends JpaRepository<PricingCategory, Long> {
-
-    Optional<PricingCategory> findByCode(String code);
-
-    @Query("from PricingCategory pc where pc.startDate <= :date and pc.endDate > :date")
-    Optional<PricingCategory> findByDate(@Param("date") LocalDate date);
-
-}
