--- registration-app/src/main/java/com/registrationapp/repository/PricingCodeRepository.java
@@ -1,11 +1,0 @@
-package com.registrationapp.repository;
-
-
-import com.registrationapp.model.PricingCategory;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PricingCodeRepository extends JpaRepository<PricingCategory, String> {
-
-}
