--- registration-app/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-liquibase.change-log=classpath:liquibase-changeLog.xml
-
