--- registration-app/src/test/java/com/registrationapp/registrationapp/RegistrationAppApplicationTests.java
@@ -1,13 +1,0 @@
-package com.registrationapp.registrationapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RegistrationAppApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
