+++ techamniac_conferention/src/main/java/com/techamniac_conferention/TechamniacConferentionApplication.java
@@ -1,0 +1,13 @@
+package com.techamniac_conferention;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TechamniacConferentionApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TechamniacConferentionApplication.class, args);
+    }
+
+}
