+++ techamniac_conferention/src/main/java/com/techamniac_conferention/model/PricingCategory.java
@@ -1,0 +1,34 @@
+package com.techamniac_conferention.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.time.LocalDate;
+import java.util.Objects;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "pricing_categories")
+public class PricingCategory {
+
+    @Id
+    @Column(name = "pricing_category_code", unique = true, nullable = false, length = 1)
+    private String code;
+
+    @Column(name = "pricing_category_name", nullable = false, length = 20)
+    private String name;
+
+    @Column(name = "pricing_start_date", nullable = false)
+    private LocalDate startDate;
+
+    @Column(name = "pricing_end_date", nullable = false)
+    private LocalDate endDate;
+
+}
