+++ techamniac_conferention/src/main/java/com/techamniac_conferention/repository/AttendeeRepository.java
@@ -1,0 +1,7 @@
+package com.techamniac_conferention.repository;
+
+import com.techamniac_conferention.model.Attendee;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AttendeeRepository extends JpaRepository<Attendee, Integer> {
+}
