+++ techamniac_conferention/src/main/java/com/techamniac_conferention/repository/TicketTypeRepository.java
@@ -1,0 +1,10 @@
+package com.techamniac_conferention.repository;
+
+import com.techamniac_conferention.model.TicketType;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
+public interface TicketTypeRepository extends JpaRepository<TicketType, String> {
+
+    Optional<TicketType> findByCode(String code);
+}
